Moved validation over to a separate file
